Load scenario via settings.json with the specified path
Scenario is now being loaded by looking into the settings.json. In it has to be the path to the "real" settings.json typically located in the Physikablgeich repository. That has to contain a "szenario_path" and "file_name" for the RLBot test exercise to find the correct scenario file